chore: move axelar tests in a single file
